% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_topic_matrices.R
\name{as.sparseMatrix}
\alias{as.sparseMatrix}
\title{Convert a \code{tidy_topic_matrix} to a sparse matrix}
\usage{
as.sparseMatrix(x)
}
\arguments{
\item{x}{a \code{\link[=as.tidy_topic_matrix]{tidy_topic_matrix}}.}
}
\description{
Converts a \code{\link[=as.tidy_topic_matrix]{tidy_topic_matrix}} to 
a sparse matrix format of \code{\link[Matrix]{sparseMatrix}} in the 
\code{Matrix} package.
}
\seealso{
\code{\link[Matrix]{sparseMatrix}}
}

