% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_io.R
\name{assert_mallet_state_file_name}
\alias{assert_mallet_state_file_name}
\title{Assert that \code{file_path} is a valide mallet state file}
\usage{
assert_mallet_state_file_name(file_path)
}
\arguments{
\item{file_path}{File path to check}
}
\description{
Assert that \code{file_path} is a valide mallet state file
}
\keyword{internal}

