% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_topic_matrices.R
\name{doc_topic_matrix}
\alias{doc_topic_matrix}
\alias{topic_doc_matrix}
\alias{topic_type_matrix}
\alias{type_topic_matrix}
\title{Tidy topic matrices}
\usage{
doc_topic_matrix(state)

topic_doc_matrix(state)

type_topic_matrix(state)

topic_type_matrix(state)
}
\arguments{
\item{state}{a \code{tidy_topic_state} object.}
}
\description{
Calculate a tidy topic document matrix, document topic matrix. 
These matrices are often very sparse.
These can easily be converted to a real sparse matrix using \code{as.sparseMatrix}.
}
\examples{
# Load the state of the union topic model
data("sotu50")
dtm <- doc_topic_matrix(sotu50)
tdm <- topic_doc_matrix(sotu50)
ttm <- type_topic_matrix(sotu50)
ttm <- topic_type_matrix(sotu50)
sp_mat <- as.sparseMatrix(x = ttm)

}
\seealso{
\code{as.sparseMatrix}
}

