% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_checking.R
\name{ggplot_imi_type}
\alias{ggplot_imi_type}
\title{Create IMI plot}
\usage{
ggplot_imi_type(observed_imi, topic, replicated_imi = NULL)
}
\arguments{
\item{observed_imi}{a \code{tbl_df} with \code{imi}, \code{type} and \code{topic} based on the original data.}

\item{topic}{which topic to plot}

\item{replicated_imi}{a \code{tbl_df} with \code{imi}, \code{type} and \code{topic} based on multiple replications. Default is NULL, no replications used.}
}
\description{
Create IMI plot
}
\details{
When sampling and deviation calculation this can be done using argument state and
and iteration number.
}

