% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.tidy_topic_array}
\alias{is.tidy_topic_array}
\alias{is.tidy_topic_matrix}
\title{Assert that \code{x} is a \code{tidy_topic_matrix} or a \code{tidy_topic_array}}
\usage{
is.tidy_topic_array(x)

is.tidy_topic_matrix(x)
}
\arguments{
\item{x}{Object to check if it is a valid \code{tidy_topic_matrix} 
or \code{tidy_topic_array}.}
}
\description{
The tidy_topic_array/tidy_topic_matrix is a \code{tibble} dataset 
with a \code{"topic"} variable and an integer variable \code{"n"}.

The \code{tidy_topic_matrix} do only contain one more variable, 
\code{tidy_topic_matrix} an array can contain multiple dimensions.

The datastructure is sparse, meaning 0 counts are not included.
}

