% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.tidy_topic_state}
\alias{assert_state}
\alias{assert_state_object}
\alias{is.tidy_topic_state}
\title{Assert that \code{x} is a tidy topic state, matrix or array}
\usage{
is.tidy_topic_state(x)

assert_state(x)

assert_state_object(x)
}
\arguments{
\item{x}{Object to check if it is a valid topic_model_state object}
}
\description{
Assert that \code{x} is a tidy topic state, matrix or array
}
\details{
The tidy topic state is a \code{tibble} dataset with at least the variables
\code{doc} (integer or factor), \code{type} (factor) and \code{topic} (integer or factor).
}
\keyword{internal}

