% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_checking.R
\name{mi_deviance}
\alias{mi_deviance}
\title{Calculate MI deviance}
\usage{
mi_deviance(true, rep)
}
\arguments{
\item{true}{a \code{tbl_df} with \code{mi} by \code{topic} based on the original data}

\item{rep}{a \code{tbl_df} with \code{mi} by \code{topic} based on multiple replications}
}
\description{
Calculate MI deviance
}
\details{
When sampling and deviation calculation this can be done using argument state and
and iteration number.
}

