% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_checking.R
\name{plot_imi_type}
\alias{plot_imi_type}
\title{Create IMI plot}
\usage{
plot_imi_type(true, rep, k)
}
\arguments{
\item{true}{a \code{tbl_df} with \code{imi}, \code{type}, \code{topic} and \code{rank} based on the original data.}

\item{rep}{a \code{tbl_df} with \code{imi}, \code{type} by \code{topic} based on multiple replications.}

\item{k}{which topic to plot}
}
\description{
Create IMI plot
}
\details{
When sampling and deviation calculation this can be done using argument state and
and iteration number.
}

