% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_io.R
\name{read_mallet_statefile}
\alias{read_mallet_statefile}
\title{Read a Mallet state file into R}
\usage{
read_mallet_statefile(state_file, type = "state")
}
\arguments{
\item{state_file}{Path to a mallet state file with \code{.gz} or \code{.txt} suffix.}

\item{type}{What to return from the state file. Can be \code{state}, \code{alpha} or \code{beta}. Default is \code{state}.}
}
\value{
The return value depends on \code{type}. 
\code{state} returns the topic indicators, word types and document id in \code{tbl_df} format.
\code{alpha} returns a vector of alpha priors values (in character format).
\code{beta} returns the beta prior (in character format).
}
\description{
This function reads a mallet file into R. the purpose is to enable the full strenght of R to visualize and postprocess sampled data from mallet
}

