% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_vanilla_lda}
\alias{sample_vanilla_lda}
\title{Sample vanilla LDA}
\usage{
sample_vanilla_lda(doc, type, z, K, D, V, iter, beta, alpha)
}
\arguments{
\item{p}{a vector of (pssoble unnormalized) probabilities.}
}
\description{
Sample a couple of iterations using simple LDA
}

