% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_documentation.R
\docType{data}
\name{sotu50}
\alias{sotu50}
\title{Topic model of State of the Union Adresses.}
\format{A \code{\link[tibble]{tibble}} \code{data.frame} with 1,263,832 rows and 3 variables:
\describe{
  \item{doc}{Document id (paragraph),}
  \item{type}{Type used.}
  \item{text}{Topic indicator.}
}}
\source{
\url{https://en.wikipedia.org/wiki/State_of_the_Union}
}
\usage{
sotu50
}
\description{
A dataset containing a topic model using 50 topics for State of the Union Adresses.
Stop words has been removed using the Mallet english stop word list.
}
\seealso{
\code{\link{sotu}}
}
\keyword{datasets}

