% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_documentation.R
\docType{data}
\name{statse}
\alias{statse}
\title{2000 question posts from stats.stackexchange.com}
\format{A \code{\link[tibble]{tibble}} \code{data.frame} with 2,000 rows and 1 variables:
\describe{
  \item{text}{Textual content in post.}
}}
\source{
\url{https://stats.stackexchange.com}
}
\usage{
statse
}
\description{
2000 question posts from stats.stackexchange.com
}
\keyword{datasets}

