% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_documentation.R
\docType{package}
\name{tidytopics}
\alias{tidytopics}
\alias{tidytopics-package}
\title{Sampling along a topic model}
\description{
#@section Implementations.
}
\details{
Vanilla LDA using Gibbs sampler.

#@section Evaluation metrics.

LogLikelihood
}
\author{
Mans Magnusson
}

