% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_summaries.R
\name{top_terms}
\alias{top_terms}
\title{Weighted top terms}
\usage{
top_terms(x, scheme = "type_probability", j = 10, beta = 0, ...)
}
\arguments{
\item{x}{A \code{\link[as.tidy_topic_state]{tidy_topic_state}}}

\item{scheme}{The weight scheme to use. Default is \code{type_probability}.}

\item{j}{The number of types to return. Default is 10.}

\item{beta}{Beta hyper parameter. Default is 0 (no prior smoothing).}

\item{...}{additional parameters used by weighting schemes. See details.}
}
\value{
Returns a \code{\link[tibble]{tibble}} with topic and top terms and weights
}
\description{
There are multiple ways to reweight topics. The following ones are implemented:

\describe{
  \item{\code{type_probability}}{\eqn{p(w|k)} The probability of a type given the topic. The most common weighting scheme.}
  \item{\code{topic_probability}}{\eqn{p(k|w)} The probability of a topic given a term.}
  \item{\code{term_score}}{\eqn{p(w|k) * log(p(w|k) / (\prod p_k(w|k))^(1/K)} A weighting scheme inspired by tf-idf proposed by Lafferty and Blei (2009).}
  \item{\code{relevance}}{\eqn{log(p(w|k)/ (\sum p_k(w)^(1-\lambda))} A weighting scheme proposed by Sievert and Shirley (2014)}
  \item{\code{n_wk}}{\eqn{n_wk} Order by number of topic indicators. Give same result as \code{type_probability} but is faster.}
}
}
\details{
Only returning values for type-topic combination that exist in the model is returned.
This means that unless \code{beta} is set to 0, the returning probabilities
will not sum to 1.

If ties in weight/probability, the original order is returned.

\code{relevance} weighting uses the additional parameter \code{lambda}. Default is 0.6.
}
\examples{
# Load the state of "the State of the Union" topic model
load(system.file("extdata/sotu50.Rdata", package = "tidytopics"))
w <- top_terms(x = sotu50, "n_wk")
w <- top_terms(x = sotu50, beta = 0.01)

}
\references{
Blei, D. M., & Lafferty, J. D. (2009). Topic models. Text mining: classification, clustering, and applications, 10(71), 34.

Sievert, C., & Shirley, K. E. (2014). LDAvis: A method for visualizing and interpreting topics. In Proceedings of the workshop on interactive language learning, visualization, and interfaces (pp. 63-70).
}

