% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_entropy.R
\name{topic_entropy}
\alias{topic_entropy}
\title{Calculate the topic entropy}
\usage{
topic_entropy(x, prior = 0)
}
\arguments{
\item{x}{a \code{tidy_topic_array}.}

\item{prior}{a smoothing constant to use (often alpha). Not currently implemented.}
}
\description{
Calculate the topic entropy
}
\details{
Calculates the entropy over topics.

Assumes that of beta is set to 0 the probability 0 is the limit when p -> 0+ i.e 0.
}
\examples{
data("sotu50")
dtm <- doc_topic_matrix(sotu50)
H_0 <- topic_entropy(x = dtm)
H_doc <- topic_entropy(x = dplyr::group_by(dtm, doc))

}

