% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_entropy.R
\name{topic_jsd}
\alias{topic_jsd}
\title{Calculate the Jensen-Shannon Distance between groups}
\usage{
topic_jsd(x, dist, prior = 0)
}
\arguments{
\item{x}{a possibly grouped \code{tidy_topic_array}.}

\item{dist}{a categorical variable to calculate JSD between group}

\item{prior}{probability smoothing/dirichlet prior. Default is 0.}
}
\description{
Calculate the Jensen-Shannon Distance between groups
}
\details{
Calculates the Jensen-Shannon distance for a grouped variable.

If prior is 0, 0 * log(0) is assumed to be 0 (ie the limit).
Assumes that of beta is set to 0 the probability 0 is the limit when p -> 0+ i.e 0.
}

