% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_entropy.R
\name{topic_probability}
\alias{topic_probability}
\title{Calculates the probability over topics}
\usage{
topic_probability(x, prior = NULL)
}
\arguments{
\item{x}{a possibly grouped \code{tidy_topic_array}.}

\item{prior}{probability smoothing/dirichlet prior. Default is NULL, no smoothing and sparse results.}
}
\description{
Calculates the probability over topics
}

