% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_io.R
\name{write.mallet.statefile}
\alias{write.mallet.statefile}
\title{Write a Mallet state from R}
\usage{
write.mallet.statefile(state.object, alpha, beta, state_file)
}
\arguments{
\item{state.object}{A data.frame in a one-token-per-row format (see \code{tidytext}).
\code{doc}: Document id as integer, 
\code{pos}: Position in the document (integer from 1 to length of document), 
\code{type}: Word type of the token as a factor variable.
\code{topic}: Topic indicator (as integer)}

\item{alpha}{A vector of length equal to the number of topics containing the alpha prior for each topic.}

\item{beta}{A numeric or character element with the beta prior.}

\item{state_file}{Path to write a mallet state file to (\code{.gz} or \code{.txt} suffix is mandatory).}
}
\value{
\code{NULL} if file was written successfully.
}
\description{
Write a Mallet state from R
}

