% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_io.R
\name{write_mallet_statefile}
\alias{write_mallet_statefile}
\title{Write a Mallet state from R}
\usage{
write_mallet_statefile(state, alpha, beta, state_file)
}
\arguments{
\item{state}{a \code{tidy_topic_state} object in a one-token-per-row format (see \code{tidytext}).}

\item{alpha}{a numeric vector of length equal one (uniform prior) or to the number of topics. This is the alpha prior for each topic.}

\item{beta}{A numeric element with the beta prior (single value).}

\item{state_file}{a character string naming a file.}
}
\value{
\code{NULL} if file was written successfully.
}
\description{
Write a Mallet state from R
}

