% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_topic_matrices.R
\name{doc_topic_matrix}
\alias{as.matrix.tidy_topic_matrix}
\alias{doc_topic_matrix}
\alias{tidy_topic_matrix}
\alias{topic_doc_matrix}
\alias{topic_type_matrix}
\alias{type_topic_matrix}
\title{Tidy topic matrices}
\usage{
doc_topic_matrix(state)

topic_doc_matrix(state)

type_topic_matrix(state)

topic_type_matrix(state)

doc_topic_matrix(state)

\method{as.matrix}{tidy_topic_matrix}(x)
}
\arguments{
\item{state}{a \code{tidy_topic_state} object.}

\item{x}{a \code{tidy_topic_matrix} object.}
}
\description{
Calculate a tidy topic matrix. These are often very sparse.
These can easily be converted to a sparse matrix using \code{as.matrix()}.
}

