% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_checking.R
\name{imi}
\alias{imi}
\alias{imi_group}
\title{Calculate IMI}
\usage{
imi(state, w = NULL)

imi_group(state, g, w = NULL)
}
\arguments{
\item{state}{a tidy_topic_state object}

\item{w}{An ungrouped tbl_df with types and topics to calculate IMI for. Default is NULL.}

\item{g}{groups TODO}
}
\description{
Function to calculate instantanueous mutual information for types for a given topic.
See reference for details.
}
\references{
Mimno, D. and Blei, D. Bayesian Checking for Topic Models
}

