% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_vanilla_lda}
\alias{sample_vanilla_lda}
\title{Sample vanilla LDA}
\usage{
sample_vanilla_lda(doc, type, z, K, D, V, iter, beta, alpha)
}
\arguments{
\item{doc}{TODO}

\item{type}{TODO}

\item{z}{TODO}

\item{K}{TODO}

\item{D}{TODO}

\item{V}{TODO}

\item{iter}{TODO}

\item{beta}{TODO}

\item{alpha}{TODO}
}
\description{
Sample a couple of iterations using simple LDA
}

