% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytopics_documentation.R
\docType{data}
\name{uspresidents}
\alias{uspresidents}
\title{US Presidents.}
\format{An object of class \code{data.frame} with 44 rows and 7 columns.}
\source{
\url{https://en.wikipedia.org/wiki/State_of_the_Union}
}
\usage{
uspresidents
}
\description{
A dataset containing US presidents from 1789 to 2015
}
\keyword{datasets}

